requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
openai==1.3.0
pydantic==2.0.0
python-dotenv==1.0.0
yt-dlp==2023.7.6
chromadb==0.4.15
pydub==0.25.1
ffmpeg-python==0.2.0
python-multipart==0.0.6
langchain==0.0.267
loguru==0.7.0
tiktoken==0.5.1
numpy==1.24.3
tenacity==8.2.3